requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
beautifulsoup4==4.12.0
black==23.3.0
certifi==2019.6.16
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
discord.py==1.7.3
idna==2.8
importlib-metadata==6.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pycparser==2.21
PyNaCl==1.4.0
requests==2.28.2
six==1.16.0
soupsieve==1.9.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.25.3
yarl==1.8.2
zipp==3.15.0
